0.1.0
Added in config/connection.js to create the connection to the mongodb with the mongodb_uri or the local connection. added in a server.js to start tje express server and initiate the mongodb connectio and the app to use the router from controllers . added a models/ which holds the individual model schemas, then they are all required inside of models/index.js where they all are exported as a group to be required with destructuring. created a seeds/index.js that can be ran with package script npm run seed, which executes a function to run all of the imported seed files, which hold a create method that match the model scehma. created a controllers/api/user.js which serves the find() method on the require model "user" and then responds with the user found in json data. router is then exported to index.js where all router instances are required and then exported to server to be utilized.