0.2.0
added /controller/user.js added a user controller object that uses async await try catch functions with req, res args, that are exported to the recently added /route/api/user.js where it is required and destructured then a router instance is defined where router is called with the dot method and the functions from controller/user as an argument. then the router is exported to index.js where it is prefixed with /api and then exported to server.js where it is imported and app.use is called on the router. all user routes complete.