0.4.0
added in additional controller functions for router instance inside /controller/thought.js added create thought, update thought, delete thought, add reaction, delete reaction, all of which use async/ await inside of try catches, and follow the methodology where a method is called on a model, where its arguments are defined with key value pairs, the key being where to target and the value being a body param or params.id then error handling is used by an if statment if an id is required and the catch statement at the end of the try block. all controlers are exported and imported/required inside of routes, where an instance of the router is called with a .route property to set the route and a . method which will take the function from the destructured require of the controller/thoughts as an arg. then inside of the route/api/thought.js the router instance is exported to the server where it is required and called with app.use to initiated by the server.